requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.0
huggingface-hub==0.19.4
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0